Update Default.aspx.cs
Sometimes we need to know page direction in css and js files. Giving a rtl class to body page makes it possible.
For example most of jquery effects like sub menu behavior are defined in JavaScript files. in this situations we can simply use a code like below:
if ($('body').hasClass('rtl')) {
    $this.css({ position: 'absolute', left: '10px' });
} else {
    $this.css({ position: 'absolute', right: '10x' });
}